backend/pyproject.toml
[project]
name = "inventoryview"
version = "0.1.0"
description = "InventoryView - Infrastructure inventory with graph-based data model"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "psycopg[binary,pool]>=3.2",
    "pyjwt[crypto]>=2.9",
    "cryptography>=44.0",
    "argon2-cffi>=23.1",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "alembic>=1.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "S",   # flake8-bandit
    "RUF", # ruff-specific rules
]
ignore = [
    "S101", # assert used (OK in tests)
    "S104", # binding to all interfaces (intentional for container)
]

[tool.ruff.lint.isort]
known-first-party = ["inventoryview"]
